requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
xgboost>=2.0
catboost>=1.2
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
pyarrow>=14.0
tqdm>=4.65
cupy-cuda12x>=13.0
cudf-cu12>=25.02
cuml-cu12>=25.02
